Refatorando os códigos das telas de CriarTreino e GerenciarTreinos com Custom Hooks
