--- components/AdicionarExerc.js
@@ -1,17 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
-
-
-
-const styles = StyleSheet.create({
-    exercInput: {
-        fontSize: 20,
-        backgroundColor: 'gray',
-        borderRadius: 5,
-        padding: 10,
-        marginBottom: 10,
-        marginHorizontal: 2,
-    },
-})
-
-export default AdicionarExerc;
