+++ package.json
@@ -13,16 +13,13 @@
     "@react-native-async-storage/async-storage": "^1.23.1",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
-    "expo": "^51.0.11",
+    "expo": "^51.0.14",
     "expo-font": "^11.10.3",
     "expo-status-bar": "^1.12.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-icons": "^5.2.1",
-    "react-native": "^0.73.8",
-    "react-native-countdown-circle-timer": "^3.2.1",
-    "react-native-countdown-component": "^2.7.1",
-    "react-native-safe-area-context": "^4.10.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-native": "^0.74.2",
+    "react-native-safe-area-context": "^4.10.4",
     "react-native-screens": "^3.31.1",
     "react-native-web": "^0.19.11"
   },
